backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
pydantic==2.4.2
pydantic-settings==2.0.3
uvicorn==0.23.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.7
alembic==1.12.0
python-dotenv==1.0.0
httpx==0.25.0